src/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pyzmq
aiokafka
opendis
pydantic-ai[openai]
openai
sse-starlette
